services/financial_data_collector/pyproject.toml
[tool.poetry]
name = "financial-data-collector"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
yfinance = "^0.2.65"
fredapi = "^0.5.2"
pandas-datareader = "^0.10.0"
pandas = "^2.3.1"
pyyaml = "^6.0.2"
sqlalchemy = "^2.0.41"
fastparquet = "^2024.11.0"
jsonschema = "^4.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
